# Optimize for strings without multibyte characters