@zenspider 2024-12-08
I've locally pulled this to resolve conflicts and the code in question has been removed entirely. I also agree with the reviewer that this wasn't a valid comparison as they didn't clear out the cached files between runs. 

Closing.